# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Badgeverso Be",
    "description": "System management s.p.a",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
  "host": "https://be-system.herokuapp.com",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/auth/forgotPassword/:email": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "password dimenticata",
        "description": "",
        "operationId": "forgotPassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "params",
            "name": "email",
            "type": "string",
            "description": "email utente",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "cambio pass",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized",
            "schema": {
              "$ref": "#/definitions/401"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "error": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "number"
        },
        "extra": {
          "type": "object"
        }
      }
    },
    "401": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "extra": {
          "type": "object"
        },
        "errors": {
          "$ref": "#/definitions/error"
        }
      }
    },
...
  }
}